+++ R/get.R
@@ -18,7 +18,7 @@
 #' \dontrun{
 #' # follow the example in https://aurin.org.au/resources/aurin-apis/aurin-open-api-and-r/
 #' aurinapi_register(username = "your-username", password = "your-password")
-#' aurinapi_get("aurin:datasource-au_govt_dss-UoM_AURIN_national_public_toilets_2017")
+#' aurinapi_get("aurin:datasource-UQ_ERG-UoM_AURIN_DB_public_toilets")
 #' }
 aurinapi_get = function(open_api_id, crs = "EPSG:4326", params = NULL) {