+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * `aurinapi_build_get_feature_request()` gains `outputFormat` argument, default as "application/json".
 * `aurinapi_search()` returns a query result from the AURIN API search.
+* Update the example data layer. `aurin:datasource-au_govt_dss-UoM_AURIN_national_public_toilets_2017` layer no longer exists hence we are replacing it with `aurin:datasource-UQ_ERG-UoM_AURIN_DB_public_toilets`, which is most similar.
 
 # aurinapi 0.3.3